chore(telemetry): use static configs
