+++ ddtrace/settings/_otel_remapper.py
@@ -8,8 +8,6 @@ from typing import Tuple
 from ..constants import ENV_KEY
 from ..constants import VERSION_KEY
 from ..internal.logger import get_logger
-from ..internal.telemetry import telemetry_writer
-from ..internal.telemetry.constants import TELEMETRY_NAMESPACE
 
 
 log = get_logger(__name__)
@@ -134,23 +132,6 @@ ENV_VAR_MAPPINGS: Dict[str, Tuple[str, Callable[[str], Optional[str]]]] = {
 }
 
 
-def validate_otel_envs():
-    user_envs = {key.upper(): value for key, value in os.environ.items()}
-    for otel_env, _ in user_envs.items():
-        if (
-            otel_env not in ENV_VAR_MAPPINGS
-            and otel_env.startswith("OTEL_")
-            and otel_env not in ("OTEL_PYTHON_CONTEXT", "OTEL_TRACES_SAMPLER_ARG", "OTEL_LOGS_EXPORTER")
-        ):
-            _unsupported_otel_config(otel_env)
-        elif otel_env == "OTEL_LOGS_EXPORTER":
-            # check for invalid values
-            otel_value = os.environ.get(otel_env, "none").lower()
-            if otel_value != "none":
-                _invalid_otel_config(otel_env)
-            telemetry_writer.add_configuration(otel_env, otel_value, "env_var")
-
-
 def parse_otel_env(otel_env: str) -> Optional[str]:
     _, otel_config_validator = ENV_VAR_MAPPINGS[otel_env]
     raw_value = os.environ.get(otel_env, "")
@@ -159,45 +140,5 @@ def parse_otel_env(otel_env: str) -> Optional[str]:
         raw_value = raw_value.lower()
     mapped_value = otel_config_validator(raw_value)
     if mapped_value is None:
-        _invalid_otel_config(otel_env)
         return None
     return mapped_value
-
-
-def hiding_otel_config(otel_env, dd_env):
-    log.debug(
-        "Datadog configuration %s is already set. OpenTelemetry configuration will be ignored: %s=%s",
-        dd_env,
-        otel_env,
-        os.environ[otel_env],
-    )
-    telemetry_writer.add_count_metric(
-        TELEMETRY_NAMESPACE.TRACERS,
-        "otel.env.hiding",
-        1,
-        (("config_opentelemetry", otel_env.lower()), ("config_datadog", dd_env.lower())),
-    )
-
-
-def _invalid_otel_config(otel_env):
-    log.warning(
-        "Setting %s to %s is not supported by ddtrace, this configuration will be ignored.",
-        otel_env,
-        os.environ.get(otel_env, ""),
-    )
-    telemetry_writer.add_count_metric(
-        TELEMETRY_NAMESPACE.TRACERS,
-        "otel.env.invalid",
-        1,
-        (("config_opentelemetry", otel_env.lower()),),
-    )
-
-
-def _unsupported_otel_config(otel_env):
-    log.warning("OpenTelemetry configuration %s is not supported by Datadog.", otel_env)
-    telemetry_writer.add_count_metric(
-        TELEMETRY_NAMESPACE.TRACERS,
-        "otel.env.unsupported",
-        1,
-        (("config_opentelemetry", otel_env.lower()),),
-    )
